Fixed an issue which could cause a segfault.
Specifically:

Changed iline 572 of packet_processing.c from:

                       arse = arse + 4 + (uint8_t)(extensions_tmp_ptr[(arse+2)]*256) + (uint8_t)(extensions_tmp_ptr[arse+3]);
to
                       arse = arse + 4 + (((uint8_t) extensions_tmp_ptr[(arse+2)])*256) + (uint8_t)(extensions_tmp_ptr[arse+3]);

Note the brackets.  I was casting extensions_tmp_ptr[(arse+2)] to a uint8_t
(8-bit number) after multiplying by 256 instead of before.  Resulting in a
number which is too small if the number was anything other than 0.  Luckily (or
terribly) this number *is* typically zero so the problem didn't manifest most
of the time, but when it does (when a single extension exceeds 256bytes in
length) then the pointer would be "wonky" and on certain systems (Linux with
gcc) would cause heap corruption and ultimately a segfault.  On OSX/BSD with
clang this issue does not cause a crash.  I can only assume that this is a
difference in implimentations of malloc, aslr, dep, or some other ram
tomfoolery.